+++ PinpadLibrary/src/main/java/io/cloudwalk/pos/pinpadlibrary/internals/commands/GTK.java
@@ -39,35 +39,35 @@ public class GTK {
         String SPE_PBKEXP   = input.getString(ABECS.SPE_PBKEXP);
 
         if (SPE_TRACKS  != null) {
-            stream[1].write(PinpadUtility.buildRequestTLV(ABECS.TYPE.N, "0007", SPE_TRACKS));
+            stream[1].write(PinpadUtility.buildTLV(ABECS.TYPE.N, "0007", SPE_TRACKS));
         }
 
         if (SPE_MTHDDAT != null) {
-            stream[1].write(PinpadUtility.buildRequestTLV(ABECS.TYPE.N, "0003", SPE_MTHDDAT));
+            stream[1].write(PinpadUtility.buildTLV(ABECS.TYPE.N, "0003", SPE_MTHDDAT));
         }
 
         if (SPE_IVCBC   != null) {
-            stream[1].write(PinpadUtility.buildRequestTLV(ABECS.TYPE.B, "001D", SPE_IVCBC));
+            stream[1].write(PinpadUtility.buildTLV(ABECS.TYPE.B, "001D", SPE_IVCBC));
         }
 
         if (SPE_OPNDIG  != null) {
-            stream[1].write(PinpadUtility.buildRequestTLV(ABECS.TYPE.N, "0008", SPE_OPNDIG));
+            stream[1].write(PinpadUtility.buildTLV(ABECS.TYPE.N, "0008", SPE_OPNDIG));
         }
 
         if (SPE_KEYIDX  != null) {
-            stream[1].write(PinpadUtility.buildRequestTLV(ABECS.TYPE.N, "0009", SPE_KEYIDX));
+            stream[1].write(PinpadUtility.buildTLV(ABECS.TYPE.N, "0009", SPE_KEYIDX));
         }
 
         if (SPE_WKENC   != null) {
-            stream[1].write(PinpadUtility.buildRequestTLV(ABECS.TYPE.N, "000A", SPE_WKENC));
+            stream[1].write(PinpadUtility.buildTLV(ABECS.TYPE.N, "000A", SPE_WKENC));
         }
 
         if (SPE_PBKMOD  != null) {
-            stream[1].write(PinpadUtility.buildRequestTLV(ABECS.TYPE.B, "0024", SPE_PBKMOD));
+            stream[1].write(PinpadUtility.buildTLV(ABECS.TYPE.B, "0024", SPE_PBKMOD));
         }
 
         if (SPE_PBKEXP  != null) {
-            stream[1].write(PinpadUtility.buildRequestTLV(ABECS.TYPE.B, "0025", SPE_PBKEXP));
+            stream[1].write(PinpadUtility.buildTLV(ABECS.TYPE.B, "0025", SPE_PBKEXP));
         }
 
         byte[] CMD_DATA = stream[1].toByteArray();