+++ DEMO/CHANGELOG.md
@@ -9,11 +9,11 @@
 - Pinpad Service 1.0.3
 
 ## [0.0.10] - 2022-06-07
-- Update internal dependencies.
+- Update dependencies.
 - Lock screen orientation (portrait).
 
 ## [0.0.9] - 2022-06-03
-- Update internal dependencies.
+- Update dependencies.
 - Disable additional local requests from previous version.
 - Review CHANGELOG.md.