+++ VirtualService/CHANGELOG.md
@@ -4,6 +4,8 @@
 - Android 5.1 (API 22)
 
 ## [0.0.3] - 2022-06-07
+- Update dependencies.
+- Intercept `GIX` response to inject `VIRTUAL//` into `PP_MODEL` value.
 - Lock screen orientation (portrait).
 
 ## [0.0.2] - 2022-02-24